requirements.txt
numpy==1.23.5
numpy==1.22.3
numpy==1.25.2
pandas==1.5.3
Pillow==10.0.0
Pillow==9.4.0
Pillow==10.0.0
streamlit==1.25.0
tabulate==0.8.10
tabulate==0.9.0
